ListingPage: add getDerivedRenderData function and start using it
